endpoint orderRating listo
